+++ lib/app_build_info/build_config.dart
@@ -1,0 +1,46 @@
+import 'dart:convert';
+
+import 'package:flutter_fit_utils/model/model.dart';
+import 'package:flutter_fit_utils_config/remote_config.dart';
+import 'package:package_info_plus/package_info_plus.dart';
+
+import '../flutter_fit_utils_vaistudio.dart';
+
+/// Configuration for the app's build information.
+class BuildConfig extends RemoteConfig {
+  /// Name of the parameter for the build information.
+  static const String versionKey = "build_version";
+
+  /// Name of the parameter for the store url.
+  static const String storeUrlKey = "store_url";
+
+  String storeUrl = "";
+
+  String currentAppVersion = "";
+  int currentBuild = 0;
+
+  BuildVersion? latestBuildInformation;
+
+  @override
+  void read() {
+    latestBuildInformation = BuildVersion.fromModel(Model.fromJson(
+        jsonDecode(appConfig.getString(versionKey)) as Map<String, dynamic>));
+
+    storeUrl = appConfig.getString(storeUrlKey);
+  }
+
+  /// Reads the current app version on the device and sends the data to the [BuildProvider].
+  Future<void> getCurrentVersion(BuildProvider buildProvider) async {
+    final PackageInfo package = await PackageInfo.fromPlatform();
+    buildProvider.currentAppVersion = package.version;
+    buildProvider.currentBuild = int.parse(package.buildNumber);
+
+    if (latestBuildInformation == null) {
+      read();
+    }
+
+    buildProvider.updateBuild(latestBuildInformation!);
+  }
+}
+
+final BuildConfig buildConfig = BuildConfig();