+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.0.1
+## 1.0.0
 
-* TODO: Describe initial release.
+- Added pre-built configurations
+- Added multiple extensions
+- Added FCM Provider
+- Added Build module
+- Added Subscription module
+- Added Notification service